#182: Implement Contact details for Opportunity profile
