+++ src/components/Dashboard/Profile/sections/ContactDetails/opportunity/index.ts
@@ -1,0 +1,5 @@
+export { OpportunityContactDetails } from "./OpportunityContactDetails";
+export {
+  createOpportunityContactDetailsSchema,
+  type OpportunityContactDetailsFormData,
+} from "./opportunityContactDetailsSchema";
