+++ src/components/Dashboard/Profile/sections/ContactDetails/opportunity/opportunityContactDetailsSchema.ts
@@ -1,0 +1,21 @@
+import { VolunteerCommunicationType } from "need4deed-sdk";
+import { z } from "zod";
+
+export const createOpportunityContactDetailsSchema = (t: (key: string) => string) => {
+  return z.object({
+    name: z.string().min(1, t("dashboard.opportunityProfile.contactDetails.validation.nameRequired")),
+    phone: z
+      .string()
+      .min(1, t("dashboard.opportunityProfile.contactDetails.validation.phoneRequired"))
+      .regex(/^\+[0-9\s]+$/, t("dashboard.opportunityProfile.contactDetails.validation.phoneInvalid")),
+    email: z
+      .string()
+      .min(1, t("dashboard.opportunityProfile.contactDetails.validation.emailRequired"))
+      .email(t("dashboard.opportunityProfile.contactDetails.validation.emailInvalid")),
+    waysToContact: z
+      .array(z.enum(Object.values(VolunteerCommunicationType)))
+      .min(1, t("dashboard.opportunityProfile.contactDetails.validation.waysToContactRequired")),
+  });
+};
+
+export type OpportunityContactDetailsFormData = z.infer<ReturnType<typeof createOpportunityContactDetailsSchema>>;
