+++ src/components/Dashboard/Profile/sections/ContactDetails/types.ts
@@ -1,0 +1,3 @@
+export type ContactDetailsRef = {
+  handleEditClick: () => void;
+};
