+++ src/components/Dashboard/Profile/sections/ContactDetails/volunteer/index.ts
@@ -1,0 +1,2 @@
+export { VolunteerContactDetails } from "./VolunteerContactDetails";
+export { createVolunteerContactDetailsSchema, type VolunteerContactDetailsFormData } from "./volunteerContactDetailsSchema";
