+++ src/config/constants.ts
@@ -9,8 +9,8 @@ export const apiPathCommunication = `/${apiPrefix}/communication`;
 export const apiPathAppreciation = `/${apiPrefix}/appreciation`;
 export const apiPathLogin = `/${apiPrefix}/auth/login`;
 export const apiPathAuthRefresh = `/${apiPrefix}/auth/refresh`;
-export const apiPathAuthEmailDomain = `${apiPrefix}/auth-email-domain/`;
-export const apiPathOpportunity = `${apiPrefix}/opportunity`;
+export const apiPathAuthEmailDomain = `/${apiPrefix}/auth-email-domain/`;
+export const apiPathOpportunity = `/${apiPrefix}/opportunity`;
 export const apiPathOption = `/${apiPrefix}/option`;
 export const apiProfilePath = `/volunteer/9`;
